refactor(status actions): move status actions to its own datasource file (#362)
* feat: make proposal workflow completely async

* feat: prepare the workflow for proposal actions

* feat: start with adding needed database tables

* add the interfaces and initial version of actions engine

* add the missing types and interfaces

* improve the status action engine

* collect all the users from different groups and prepare for email sending

* add sending of emails and improve the code readability

* feat: execute status actions after proposal status change

* fix linting issue

* imrpove queries and fix some issues with workflowId not received on getProposal

* fix: Every call should have a workflow attached to it

* fix empty actions reset

* move rabbitmq placeholder handler into its own file

* feat: add status actions UI in the workflows

* add description and insert some status actions

* refactor and put the config in its own component

* start connecting UI with the endpoints

* connect UI with the newly created endpoints

* improve connection between status actions UI and backend

* finalize connection between the status actions UI and backend endpoints

* final UI touches

* finalize design and UI changes

* remove default_config because it is not needed for now

* add status action validation

* test: add e2e tests for status actions feature

* update package-lock.json

* finalize status actions e2e tests

* feat: query email templates from sparkpost api

* feat: add possibility to send emails using sink_email in other environments

* refactor(status actions): move status actions to its own datasource file

* fix the check for production environment

* add timeout option for wait-on so it stops infinite polling

* fix e2e dependency config

* fix stfc dependency config

* remove status actions redundant code
