+++ apps/backend/package.json
@@ -18,8 +18,8 @@
     "tsc": "tsc",
     "build": "rm -rf ./build && tsc",
     "dev": "docker-compose up -d --remove-orphans && ts-node-dev -r tsconfig-paths/register --respawn -T ./index.ts",
-    "dev:e2e": "docker-compose -f docker-compose.e2e.yml up -d --remove-orphans && wait-on --verbose --interval 5000 --delay 15000 tcp:5432 && ts-node-dev -r tsconfig-paths/register --respawn -T ./index.ts",
-    "dev:e2e:stfc": "docker-compose -f docker-compose-stfc.e2e.yml up -d --remove-orphans && wait-on --verbose --interval 5000 --delay 15000 tcp:5432 && ts-node-dev -r tsconfig-paths/register --respawn -T ./index.ts",
+    "dev:e2e": "docker-compose -f docker-compose.e2e.yml up -d --remove-orphans && wait-on --verbose --interval 5000 --delay 15000 --timeout 300000 tcp:5432 && ts-node-dev -r tsconfig-paths/register --respawn -T ./index.ts",
+    "dev:e2e:stfc": "docker-compose -f docker-compose-stfc.e2e.yml up -d --remove-orphans && wait-on --verbose --interval 5000 --delay 15000 --timeout 300000 tcp:5432 && ts-node-dev -r tsconfig-paths/register --respawn -T ./index.ts",
     "dev:docker": "npm install && ts-node-dev --inspect=0.0.0.0:9229 --respawn -T ./index.ts",
     "prod": "tsc && node ./build/index.js",
     "lint": "tsc --noEmit && eslint . --ext .js,.ts --quiet",
