+++ apps/backend/src/config/dependencyConfigTest.ts
@@ -22,6 +22,7 @@ import { SampleDataSourceMock } from '../datasources/mockups/SampleDataSource';
 import ScheduledEventDataSourceMock from '../datasources/mockups/ScheduledEventDataSource';
 import { SEPDataSourceMock } from '../datasources/mockups/SEPDataSource';
 import { ShipmentDataSourceMock } from '../datasources/mockups/ShipmentDataSource';
+import { StatusActionsDataSourceMock } from '../datasources/mockups/StatusActionsDataSource';
 import SystemDataSourceMock from '../datasources/mockups/SystemDataSource';
 import { TemplateDataSourceMock } from '../datasources/mockups/TemplateDataSource';
 import { UnitDataSourceMock } from '../datasources/mockups/UnitDataSource';
@@ -52,6 +53,7 @@ mapClass(Tokens.PdfTemplateDataSource, PdfTemplateDataSourceMock);
 mapClass(Tokens.ProposalDataSource, ProposalDataSourceMock);
 mapClass(Tokens.ProposalEsiDataSource, ProposalEsiDataSourceMock);
 mapClass(Tokens.ProposalSettingsDataSource, ProposalSettingsDataSourceMock);
+mapClass(Tokens.StatusActionsDataSource, StatusActionsDataSourceMock);
 mapClass(Tokens.QuestionaryDataSource, QuestionaryDataSourceMock);
 mapClass(Tokens.RedeemCodesDataSource, RedeemDataSourceMock);
 mapClass(Tokens.ReviewDataSource, ReviewDataSourceMock);
