+++ apps/backend/src/queries/ProposalSettingsQueries.ts
@@ -3,6 +3,7 @@ import { inject, injectable } from 'tsyringe';
 
 import { Tokens } from '../config/Tokens';
 import { ProposalSettingsDataSource } from '../datasources/ProposalSettingsDataSource';
+import { StatusActionsDataSource } from '../datasources/StatusActionsDataSource';
 import { Authorized } from '../decorators';
 import { MailService } from '../eventHandlers/MailService/MailService';
 import { Event, EventLabel } from '../events/event.enum';
@@ -23,6 +24,8 @@ export default class ProposalSettingsQueries {
   constructor(
     @inject(Tokens.ProposalSettingsDataSource)
     public dataSource: ProposalSettingsDataSource,
+    @inject(Tokens.StatusActionsDataSource)
+    public statusActionsDataSource: StatusActionsDataSource,
     @inject(Tokens.MailService)
     public emailService: MailService
   ) {}
@@ -142,14 +145,16 @@ export default class ProposalSettingsQueries {
 
   @Authorized([Roles.USER_OFFICER])
   async getStatusAction(agent: UserWithRole | null, actionId: number) {
-    const statusAction = await this.dataSource.getStatusAction(actionId);
+    const statusAction = await this.statusActionsDataSource.getStatusAction(
+      actionId
+    );
 
     return statusAction;
   }
 
   @Authorized([Roles.USER_OFFICER])
   async getStatusActions(agent: UserWithRole | null) {
-    const statusActions = await this.dataSource.getStatusActions();
+    const statusActions = await this.statusActionsDataSource.getStatusActions();
 
     return statusActions;
   }
@@ -159,7 +164,10 @@ export default class ProposalSettingsQueries {
     agent: UserWithRole | null,
     { connectionId, workflowId }: { connectionId: number; workflowId: number }
   ) {
-    return this.dataSource.getConnectionStatusActions(connectionId, workflowId);
+    return this.statusActionsDataSource.getConnectionStatusActions(
+      connectionId,
+      workflowId
+    );
   }
 
   @Authorized([Roles.USER_OFFICER])
