+++ apps/backend/src/statusActionEngine/emailActionHandler.ts
@@ -2,7 +2,7 @@ import { logger } from '@user-office-software/duo-logger';
 import { container } from 'tsyringe';
 
 import { Tokens } from '../config/Tokens';
-import { ProposalSettingsDataSource } from '../datasources/ProposalSettingsDataSource';
+import { StatusActionsDataSource } from '../datasources/StatusActionsDataSource';
 import { MailService } from '../eventHandlers/MailService/MailService';
 import { ConnectionHasStatusAction } from '../models/ProposalStatusAction';
 import {
@@ -87,10 +87,11 @@ export const emailActionHandler = async (
 const markStatusActionAsExecuted = async (
   proposalStatusAction: ConnectionHasStatusAction
 ) => {
-  const proposalSettingsDataSource: ProposalSettingsDataSource =
-    container.resolve(Tokens.ProposalSettingsDataSource);
+  const statusActionsDataSource: StatusActionsDataSource = container.resolve(
+    Tokens.StatusActionsDataSource
+  );
 
-  await proposalSettingsDataSource.updateConnectionStatusAction({
+  await statusActionsDataSource.updateConnectionStatusAction({
     ...proposalStatusAction,
     executed: true,
   });
